Remove various debug-trace instances
